+++ packages/svelte/src/icons/CircleMinus.svelte
@@ -18,13 +18,6 @@
   stroke-linejoin="round"
   {...$$restProps}
 >
-  <g clip-path="url(#clip0_3387_114)">
-<circle cx="12" cy="12" r="9" stroke="black" stroke-width="2"/>
-<path d="M9 13C8.44772 13 8 12.5523 8 12V12C8 11.4477 8.44772 11 9 11H15C15.5523 11 16 11.4477 16 12V12C16 12.5523 15.5523 13 15 13H9Z" fill="#141313"/>
-</g>
-<defs>
-<clipPath id="clip0_3387_114">
-<rect width="24" height="24" fill="white"/>
-</clipPath>
-</defs>
+  <circle cx="12" cy="12" r="9" stroke="black" stroke-width="2"/>
+<path d="M9 13C8.44772 13 8 12.5523 8 12C8 11.4477 8.44772 11 9 11H15C15.5523 11 16 11.4477 16 12C16 12.5523 15.5523 13 15 13H9Z" fill="black"/>
 </svg>