+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @melospot/icons-react
 
+## 0.0.5
+
+### Patch Changes
+
+- Fixing the stroke fill issue on icons
+- Updated dependencies
+  - @melospot/icons@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes